Move everything to inner namespaces
