+++ reflex/components/base/__init__.py
@@ -7,3 +7,6 @@ from .head import Head
 from .link import RawLink, ScriptTag
 from .meta import Description, Image, Meta, Title
 from .script import Script
+
+fragment = Fragment.create
+script = Script.create
