+++ reflex/components/datadisplay/__init__.py
@@ -1,0 +1,6 @@
+"""Data grid components."""
+
+from .dataeditor import DataEditor, DataEditorTheme
+
+data_editor = DataEditor.create
+data_editor_theme = DataEditorTheme
