+++ reflex/components/gridjs/__init__.py
@@ -1,0 +1,5 @@
+"""Grid components."""
+
+from .datatable import DataTable
+
+data_table = DataTable.create
