+++ reflex/components/markdown/__init__.py
@@ -1,3 +1,5 @@
 """Markdown components."""
 
 from .markdown import Markdown
+
+markdown = Markdown.create
