+++ reflex/components/moment/__init__.py
@@ -1,0 +1,5 @@
+"""Moment.js component."""
+
+from .moment import Moment
+
+moment = Moment.create
