+++ reflex/components/next/__init__.py
@@ -2,7 +2,9 @@
 
 from .base import NextComponent
 from .image import Image
+from .link import NextLink
 from .video import Video
 
 image = Image.create
 video = Video.create
+next_link = NextLink.create
