+++ reflex/components/plotly/__init__.py
@@ -1,0 +1,5 @@
+"""Plotly components."""
+
+from .plotly import Plotly
+
+plotly = Plotly.create
