+++ reflex/components/react_player/__init__.py
@@ -1,3 +1,7 @@
 """React Player component for audio and video."""
 
-from .react_player import ReactPlayer
+from .audio import Audio
+from .video import Video
+
+audio = Audio.create
+video = Video.create
