+++ reflex/components/react_player/audio.pyi
@@ -7,7 +7,7 @@ from typing import Any, Dict, Literal, Optional, Union, overload
 from reflex.vars import Var, BaseVar, ComputedVar
 from reflex.event import EventChain, EventHandler, EventSpec
 from reflex.style import Style
-from reflex.components.react_player import ReactPlayer
+from reflex.components.react_player.react_player import ReactPlayer
 
 class Audio(ReactPlayer):
     pass
