+++ reflex/components/react_player/video.py
@@ -1,5 +1,5 @@
 """A video component."""
-from reflex.components.react_player import ReactPlayer
+from reflex.components.react_player.react_player import ReactPlayer
 
 
 class Video(ReactPlayer):
